Move pointsIO.go
