+++ encoding/kzgrs/verifier/verifier.go
@@ -11,7 +11,6 @@ import (
 	enc "github.com/Layr-Labs/eigenda/encoding"
 	"github.com/Layr-Labs/eigenda/encoding/kzgrs"
 	"github.com/Layr-Labs/eigenda/encoding/rs"
-	"github.com/Layr-Labs/eigenda/encoding/utils"
 	kzg "github.com/Layr-Labs/eigenda/pkg/kzg"
 	bls "github.com/Layr-Labs/eigenda/pkg/kzg/bn254"
 	wbls "github.com/Layr-Labs/eigenda/pkg/kzg/bn254"
@@ -34,7 +33,7 @@ func NewVerifier(config *kzgrs.KzgConfig, loadG2Points bool) (*Verifier, error) 
 	}
 
 	// read the whole order, and treat it as entire SRS for low degree proof
-	s1, err := utils.ReadG1Points(config.G1Path, config.SRSNumberToLoad, config.NumWorker)
+	s1, err := kzgrs.ReadG1Points(config.G1Path, config.SRSNumberToLoad, config.NumWorker)
 	if err != nil {
 		log.Println("failed to read G1 points", err)
 		return nil, err
@@ -49,13 +48,13 @@ func NewVerifier(config *kzgrs.KzgConfig, loadG2Points bool) (*Verifier, error) 
 			return nil, fmt.Errorf("G2Path is empty. However, object needs to load G2Points")
 		}
 
-		s2, err = utils.ReadG2Points(config.G2Path, config.SRSNumberToLoad, config.NumWorker)
+		s2, err = kzgrs.ReadG2Points(config.G2Path, config.SRSNumberToLoad, config.NumWorker)
 		if err != nil {
 			log.Println("failed to read G2 points", err)
 			return nil, err
 		}
 
-		g2Trailing, err = utils.ReadG2PointSection(
+		g2Trailing, err = kzgrs.ReadG2PointSection(
 			config.G2Path,
 			config.SRSOrder-config.SRSNumberToLoad,
 			config.SRSOrder, // last exclusive
@@ -155,7 +154,7 @@ func (g *Verifier) newKzgVerifier(params rs.EncodingParams) (*ParametrizedVerifi
 // we leave it as a method of the KzgEncoderGroup
 func (v *Verifier) VerifyCommit(lengthCommit *wbls.G2Point, lowDegreeProof *wbls.G2Point, length uint64) error {
 
-	g1Challenge, err := utils.ReadG1Point(v.SRSOrder-length, v.KzgConfig)
+	g1Challenge, err := kzgrs.ReadG1Point(v.SRSOrder-length, v.KzgConfig)
 	if err != nil {
 		return err
 	}
@@ -187,7 +186,7 @@ func (v *ParametrizedVerifier) VerifyFrame(commit *wbls.G1Point, f *enc.Frame, i
 		return err
 	}
 
-	g2Atn, err := utils.ReadG2Point(uint64(len(f.Coeffs)), v.KzgConfig)
+	g2Atn, err := kzgrs.ReadG2Point(uint64(len(f.Coeffs)), v.KzgConfig)
 	if err != nil {
 		return err
 	}
